+++ skrl/models/torch/deterministic.py
@@ -15,7 +15,9 @@ class DeterministicMixin:
         :param role: Role played by the model.
         """
         self._d_clip_actions = clip_actions
-        self._d_clip_actions_min, self._d_clip_actions_max = compute_space_limits(self.action_space, device=self.device)
+        self._d_min_actions, self._d_max_actions = compute_space_limits(
+            self.action_space, device=self.device, none_if_unbounded="any"
+        )
 
     def act(self, inputs: dict[str, Any], *, role: str = "") -> tuple[torch.Tensor, dict[str, Any]]:
         """Act deterministically in response to the observations/states of the environment.
@@ -35,6 +37,6 @@ class DeterministicMixin:
 
         # clip actions
         if self._d_clip_actions:
-            actions = torch.clamp(actions, min=self._d_clip_actions_min, max=self._d_clip_actions_max)
+            actions = torch.clamp(actions, min=self._d_min_actions, max=self._d_max_actions)
 
         return actions, outputs