+++ skrl/utils/runner/warp/runner.py
@@ -7,14 +7,15 @@ import dataclasses
 import math  # noqa
 
 from skrl import logger
-from skrl.agents.torch import Agent
-from skrl.envs.wrappers.torch import MultiAgentEnvWrapper, Wrapper
-from skrl.models.torch import Model
-from skrl.resources.noises.torch import GaussianNoise, OrnsteinUhlenbeckNoise  # noqa
-from skrl.resources.preprocessors.torch import RunningStandardScaler  # noqa
-from skrl.resources.schedulers.torch import KLAdaptiveLR  # noqa
-from skrl.trainers.torch import Trainer
+from skrl.agents.warp import Agent
+from skrl.envs.wrappers.warp import MultiAgentEnvWrapper, Wrapper
+from skrl.models.warp import Model
+from skrl.resources.noises.warp import GaussianNoise, OrnsteinUhlenbeckNoise  # noqa
+from skrl.resources.preprocessors.warp import RunningStandardScaler  # noqa
+from skrl.resources.schedulers.warp import KLAdaptiveLR  # noqa
+from skrl.trainers.warp import Trainer
 from skrl.utils import set_seed
+from skrl.utils.framework.warp import scalar_mul
 
 
 class Runner:
@@ -74,68 +75,27 @@ class Runner:
 
         :return: skrl component.
         """
-        from skrl.agents.torch.a2c import A2C, A2C_CFG
-        from skrl.agents.torch.amp import AMP, AMP_CFG
-        from skrl.agents.torch.cem import CEM, CEM_CFG
-        from skrl.agents.torch.ddpg import DDPG, DDPG_CFG
-        from skrl.agents.torch.ddqn import DDQN, DDQN_CFG
-        from skrl.agents.torch.dqn import DQN, DQN_CFG
-        from skrl.agents.torch.ppo import PPO, PPO_CFG
-        from skrl.agents.torch.rpo import RPO, RPO_CFG
-        from skrl.agents.torch.sac import SAC, SAC_CFG
-        from skrl.agents.torch.td3 import TD3, TD3_CFG
-        from skrl.agents.torch.trpo import TRPO, TRPO_CFG
-        from skrl.memories.torch import RandomMemory
-        from skrl.multi_agents.torch.ippo import IPPO, IPPO_CFG
-        from skrl.multi_agents.torch.mappo import MAPPO, MAPPO_CFG
-        from skrl.trainers.torch import SequentialTrainer, SequentialTrainerCfg
-        from skrl.utils.model_instantiators.torch import (
-            categorical_model,
-            deterministic_model,
-            gaussian_model,
-            multicategorical_model,
-            multivariate_gaussian_model,
-            shared_model,
-        )
+        from skrl.agents.warp.ddpg import DDPG, DDPG_CFG
+        from skrl.agents.warp.ppo import PPO, PPO_CFG
+        from skrl.agents.warp.sac import SAC, SAC_CFG
+        from skrl.memories.warp import RandomMemory
+        from skrl.trainers.warp import SequentialTrainer, SequentialTrainerCfg
+        from skrl.utils.model_instantiators.warp import deterministic_model, gaussian_model, shared_model
 
         component = {
             # models
             "gaussianmixin": gaussian_model,
-            "categoricalmixin": categorical_model,
-            "multicategoricalmixin": multicategorical_model,
             "deterministicmixin": deterministic_model,
-            "multivariategaussianmixin": multivariate_gaussian_model,
             "shared": shared_model,
             # memories
             "randommemory": RandomMemory,
             # agents
-            "a2c": A2C,
-            "a2c_cfg": A2C_CFG,
-            "amp": AMP,
-            "amp_cfg": AMP_CFG,
-            "cem": CEM,
-            "cem_cfg": CEM_CFG,
             "ddpg": DDPG,
             "ddpg_cfg": DDPG_CFG,
-            "ddqn": DDQN,
-            "ddqn_cfg": DDQN_CFG,
-            "dqn": DQN,
-            "dqn_cfg": DQN_CFG,
             "ppo": PPO,
             "ppo_cfg": PPO_CFG,
-            "rpo": RPO,
-            "rpo_cfg": RPO_CFG,
             "sac": SAC,
             "sac_cfg": SAC_CFG,
-            "td3": TD3,
-            "td3_cfg": TD3_CFG,
-            "trpo": TRPO,
-            "trpo_cfg": TRPO_CFG,
-            # multi-agents
-            "ippo": IPPO,
-            "ippo_cfg": IPPO_CFG,
-            "mappo": MAPPO,
-            "mappo_cfg": MAPPO_CFG,
             # trainers
             "sequentialtrainer": SequentialTrainer,
             "sequentialtrainer_cfg": SequentialTrainerCfg,
@@ -185,7 +145,7 @@ class Runner:
         if "rewards_shaper_scale" in cfg:
             scale = cfg["rewards_shaper_scale"]
             if scale is not None and scale != 1.0:
-                cfg["rewards_shaper"] = lambda rewards, *args, **kwargs: rewards * scale
+                cfg["rewards_shaper"] = lambda rewards, *args, **kwargs: scalar_mul(rewards, scale)
             del cfg["rewards_shaper_scale"]
 
         # backward compatibility
@@ -363,6 +323,10 @@ class Runner:
         state_spaces = env.state_spaces if multi_agent else {"agent": env.state_space}
         action_spaces = env.action_spaces if multi_agent else {"agent": env.action_space}
 
+        # override cfg
+        if cfg.get("agent", {}).get("mixed_precision") is not None:
+            del cfg["agent"]["mixed_precision"]  # mixed precision is not supported in Warp
+
         # get agent class
         if "agent" not in cfg:
             raise ValueError(f"The 'agent' field is not defined in the specified configuration")
@@ -385,50 +349,7 @@ class Runner:
         }
 
         # single-agent configuration and instantiation
-        if agent_class in ["amp"]:
-            agent_id = possible_agents[0]
-            try:
-                amp_observation_space = env.amp_observation_space
-            except Exception as e:
-                logger.warning(
-                    "Unable to get AMP space via 'env.amp_observation_space'. Using 'env.observation_space' instead"
-                )
-                amp_observation_space = observation_spaces[agent_id]
-            agent_cfg = dataclasses.asdict(self._component(f"{agent_class}_CFG")(**self._process_cfg(cfg["agent"])))
-            agent_cfg.get("observation_preprocessor_kwargs", {}).update(
-                {"size": observation_spaces[agent_id], "device": device}
-            )
-            agent_cfg.get("state_preprocessor_kwargs", {}).update({"size": state_spaces[agent_id], "device": device})
-            agent_cfg.get("value_preprocessor_kwargs", {}).update({"size": 1, "device": device})
-            agent_cfg.get("amp_observation_preprocessor_kwargs", {}).update(
-                {"size": amp_observation_space, "device": device}
-            )
-
-            motion_dataset = None
-            if "motion_dataset" in cfg:
-                if "class" not in cfg["motion_dataset"]:
-                    raise ValueError(f"The 'motion_dataset.class' field is not defined in the specified configuration")
-                motion_dataset_class = self._component(cfg["motion_dataset"]["class"])
-                motion_dataset = motion_dataset_class(device=device, **self._process_cfg(cfg["motion_dataset"]))
-            reply_buffer = None
-            if "reply_buffer" in cfg:
-                if "class" not in cfg["reply_buffer"]:
-                    raise ValueError(f"The 'reply_buffer.class' field is not defined in the specified configuration")
-                reply_buffer_class = self._component(cfg["reply_buffer"]["class"])
-                reply_buffer = reply_buffer_class(device=device, **self._process_cfg(cfg["reply_buffer"]))
-
-            agent_kwargs = {
-                "models": models[agent_id],
-                "memory": memories[agent_id],
-                "observation_space": observation_spaces[agent_id],
-                "state_space": state_spaces[agent_id],
-                "action_space": action_spaces[agent_id],
-                "amp_observation_space": amp_observation_space,
-                "motion_dataset": motion_dataset,
-                "reply_buffer": reply_buffer,
-                "collect_reference_motions": lambda num_samples: env.collect_reference_motions(num_samples),
-            }
-        elif agent_class in ["a2c", "cem", "ddpg", "ddqn", "dqn", "ppo", "rpo", "sac", "td3", "trpo"]:
+        if agent_class in ["ddpg", "ppo", "sac"]:
             agent_id = possible_agents[0]
             agent_cfg = dataclasses.asdict(self._component(f"{agent_class}_CFG")(**self._process_cfg(cfg["agent"])))
             agent_cfg.get("observation_preprocessor_kwargs", {}).update(
@@ -447,22 +368,7 @@ class Runner:
             }
         # multi-agent configuration and instantiation
         elif agent_class in ["ippo", "mappo"]:
-            agent_cfg = dataclasses.asdict(self._component(f"{agent_class}_CFG")(**self._process_cfg(cfg["agent"])))
-            agent_cfg.get("observation_preprocessor_kwargs", {}).update(
-                {agent_id: {"size": observation_spaces[agent_id], "device": device} for agent_id in possible_agents}
-            )
-            agent_cfg.get("state_preprocessor_kwargs", {}).update(
-                {agent_id: {"size": state_spaces[agent_id], "device": device} for agent_id in possible_agents}
-            )
-            agent_cfg.get("value_preprocessor_kwargs", {}).update({"size": 1, "device": device})
-            agent_kwargs = {
-                "models": models,
-                "memories": memories,
-                "observation_spaces": observation_spaces,
-                "state_spaces": state_spaces,
-                "action_spaces": action_spaces,
-                "possible_agents": possible_agents,
-            }
+            raise NotImplementedError
         return self._component(agent_class)(cfg=agent_cfg, device=device, **agent_kwargs)
 
     def _generate_trainer(self, env: Wrapper | MultiAgentEnvWrapper, cfg: dict[str, Any], agent: Agent) -> Trainer: